Add fields for WIT validation
When verifying a work item changes validity, we get a coarse interface
that determine the work item is not valid and which fields are not
valid, but not why (other than the field is required and we have not
supplied a value). However, there may be (and often are) other reasons
why the WIT edit is invalid. This change adds three new properties to
IField to aid Agents and debugging when saving invalid work items. Three
new properties are: IsDirty, IsEditable, and ValidationState. Each field
is populated from the underlying TFS WorkItem.

- *IsDirty* indicates the field has been changed, either by the system
or by the client.
- *IsEditable* indicates the field is editable by the client.
- *ValidationState* states the reason the field may not be valid:
invalid characters, invalid computed value, invalid date, invalid
because it's empty, invalid because it is empty or has an old value, an
invalid format, invalid list value, invalid because it's NOT empty,
invalid because it's not the old value, invalid because IT IS the old
value, invalid path (e.g. area or iteration), invalid because it's too
long, invalid for the expected type, invalid because the value conflicts
with another field, invalid because a value is missing in another field
(e.g. product and release), invalid for an unknown reason, and finally
that it is valid.

The `ValidationState` will assist agents at determining if there's
something that can be done to correct the issue prior to submitting the
WIT for saving. The `IsEditable` will prevent us from making changes
that are invalid.

For example:
the Release field must have a value set conditionally on the product
field. If the product is "Future", then no value is permitted in the
release field unless one previously existed.

Due to the validation rules in Agents, when a Release field is
encountered that is empty, the value "TH2" is inserted by default.  In
this case where the product has a value of "Future", the release field
has the IsEditable value of false, as the WIT cannot be assigned to a
specific release. When the WIT is validated, it is said to be invalid.
With the new properties, we can determine that the field is dirty, is
not editable, and the validation state is InvalidNotEmpty. Agents can
now intelligently adjust behavior, in this case performing the function
suggested by the validation state property, setting the value to empty.