Add support for conditional type parsing
It is not necessary to always invoke the type parser. The occasion for which it is necessary is when the TFS model stores a value as one type, and our object requires it in another.

Example: TFS stores values in "Custom String 0x" fields as System.String, but we often store decimal values.