Update CredentialsFactory
It is often the case that it is unknown which type of credential the server may accept. The `CredentialsFactory` is updated to build a collection of credentials given an optional username and password. If no password is specified the default (network) credentials are attempted, then the logged on user of the process.

This supports all versions of authentication except OAuth with TFS.

+semver: breaking