Correct issue with work item edit identity
When editing a work item, the TfsTeamProjectCollection would authorize the changes as the current interactive Windows user, not the user with the supplied credentials. This caused unexpected values to show up in work item history (having work items edited by a developer), or in environments where no windows user is logged on (like Azure hosted), an access denied fault.