Treat .ToString() as a cast when query rewriting
.ToString() in a WIQL query has no effect, so rather than throwing a
NotSupportedException, simply treat it as a cast and continue. Added a
unit test to verify correct translation.

Scenario: Consider a work item type that has a mapped field of type
`Object`. When querying on that field in LINQ (e.g. `.Where(item =>
item.ObjectField == "SomeValue")`) the ReSharper will emit a warning as this
is a reference equality check (since `ObjectField` is of type
`Object`) and would usually be a bug. The "correction" suggested is to
change the expression to `item.ObjectField.ToString() == "SomeValue"`.

Since we're translating to WIQL and WIQL will never do reference equality
anyways, this is safe and allows consumers to write clean code.