Add Indexer support to Qwiq.Linq
Add support for property indexers in Qwiq.Linq. We accomplish this by
adding an `IndexerExpression`, then treating the value of the indexer as a
member. For example, a query like this:

    `item => item["MyProperty"] == "My Value"`

will be treated as if it is written as:

    `item => item.MyProperty == "My Value"`

The `IFieldMapper` is then responsible for doing any additional property
-> name mapping if needed.