DRY-out square brackets in field names
The `FieldMapper` returned each field name in square brackets (e.g. '[Work
Item Type]') for use in WIQL. This means that users that implement their
own `IFieldMapper` need to do the same, and in multiple places.

Instead we can centralize the logic in `MemberFragment`, which is
specificially for this purpose.