Add support for null substitution
Extend the `FieldDefinitionAttribute` to allow for a value to be substituted when the source field is empty.

For example, if a non-compulsory field "Priority" is Int32, and the target property type is also Int32, a default value of 0 is placed on the target.

This permits users to create properties with one of two types: Nullable<Int32> or Int32 with the substitute.

Resolves #112

+semver: minor