Update Benchmarks
Benchmarks are now separated into two namespaces: the actual benchmark and a test that exercises the benchmark code path. The former cannot be executed under Debug and uses a special test context to enforce constraints. The latter uses the standard test context and exercises the benchmark code path to ensure it remains functional.