Add integration tests for known field behavior differences
Tests for Core fields are broken up into three types:
 - Core fields with no known differences
 - Core fields with identities
 - Core fields with date/time

The core fields with no known differences make regular equality assertions between work items produced by each client. It is expected that there are no known differences.

The core fields with identities may have differences between each client: SOAP will produce an identity in either the "Display Name" or "Display Name <account>" depending on ambiguity. REST always produces the latter.  Identities are pushed through a helper class (`IdentityFieldValue`) to parse out the different pieces and compare the resulting `DisplayName` for equality. It is expected that there are no known differences.

The core fields with date/time have differences: SOAP converts the value to the offset provided in the user's preferences, while REST always sends the value as UTC.  Each date value is returned and SOAP values are converted to UTC via the built-in method on the `DateTime` type for equality. It is expected that there are no known differences.