Intern string names when creating field definition
NET automatically performs string interning for all string literals. This is done by means of an intern pool – a special table that stores references to all unique strings. However, only string literals are interned during the compile stage, so strings created at runtime are not checked for already being added to the pool.

This is circumvented by adding the strings to the pool directly at runtime via `String.Intern`. The trade off is that the strings remain in the intern pool (and memory) for the lifetime of the `AppDomain`.