Correct NullReferenceException issues with links
When querying for a link, a value may be returned with a source of null, target of an item, and a link of null.

In SOAP: 0, 12345, 0 respectively
In Rest: NULL, 12345, NULL respectively

This represents a "self" link, indicating the value is a root (e.g. has no parents) in a recursive or hierarchy query. However, when attempting to evaluate the link type lazily either by calling the `LinkType` property or `ToString` method, a NullReferenceException was thrown because the value of the property is also null.

This was not noticed before because items were never under test or the `LinkType` property was never called. Since the value was always evaluated lazily, items ran through the system without issue.

Since this is a valid link and is used to indicate a "root" when composing a hierarchy, but is not a valid link type, the logic has been updated to permit error-free execution in eager load scenarios.

Tests have been added to support the various hierarchy scenarios with various runtime flags to execute both eager and lazy code paths.