Correct issue with Revision
SOAP client had its own interface for revisions, causing the property `Revisions` to not be overriden on the `WorkItem` base class.  Since the property is not implemented by default, the WebJobs integration project failed.

The interface has been corrected and a new SOAP integration test added to verify basic functionality of the scenario.