Correct NullReferenceException during REST WorkItemStore disposal
In some circumstances the native work item store client may not have been created. During `Dispose`, the value is attempted to be cleaned up, which may cause a `NullReferenceException` to be encountered.

Null checks have been added to Dispose method to verify state prior to making a disposal request and value assignment.