Update type conversion to prevent FirstChanceExceptions
Update implementations of `object ParseImpl(Type, object)` and `object ParseImpl(Type, object, object)` to get the underlying type code for the destination and perform special checks against the value to be converted. If the type code is one that is known to not accept an empty string and the value to be converted is an empty string, then either get the default value of the destination type or return the supplied default value.

+semver: patch