When building attribute cache, use field for WorkItemType rather than WorkItemType.Name property
The new `IWorkItem` interface has a property named `WorkItemType`, which contains the name of the work item type for that work item. This is equivalent to the property `IWorkItem.Type.Name`, except that the work item store does not receive a query to load WIT meta data for just the name.