Provide better error message when unexpected values encountered for WIT
In order to load the `Type` property on `WorkItem`, the `System.TeamProject` and `System.WorkItemType` fields must have values. Additionally, checks are performed against the project and work item types collections to verify a project and WIT are present, avoiding another KeyNotFoundException at runtime. The changes are targeted to the REST client because the SOAP client will lazily retrieve the fields when asked whereas REST will not.

This change includes new integration tests to verify the functionality.