Update AttributeMapperStrategy to not require WIT.Fields
When building the property cache for a work item type and model, the attribute mapper would verify the decorated property/field existed as a field in the WIT. This is not necessary, as an invalid field would return `null`. In REST, we use the object accessor to get the initial field value, which would try to load the WIT and fail.

`WorkItemCore` has been updated to catch the `InvalidOperationException` thrown by the REST client and attempt to read values from that point forward using the dictionary supplied by the REST API.