Separate property annotation validation
Currently the `AttributeMapperStrategy` concerns itself with providing mapping from `IWorkItem` to an arbitrary type in addition to validating if the properties are valid for such a mapping. The validation concern is now externalized through an interface, allowing callers to control the behavior of the validation. The default validation is equal to behavior introduced in 8.1.0