Increase field validation check during mapping (#168)
* Add paging ability for projects

When there are more than 100 projects in a collection, only 100 are loaded.

Resolves #162

* Add more rigorous check for valid field

This reverts to behavior prior to 8.0.0, which for REST clients, incurs a minimum 500ms penalty when hydrating `IWorkItem` instances due to the lookup time of all projects and all WITs for a referenced project. At minimum it is one call to load all projects (this is relatively fast), and one call for all WITs in that project (this can be slow). This penalty is observed once per lifetime of `IWorkItemStore`.

However, this change also prevents instances of `DeniedOrNotExistException` and `AttributeMapException` to be thrown during mapping operations, which may occur in the hundreds to thousands depending on mapping configuration.