Fix: The REST model contains duplicate field definitions
The REST model contains duplicates for all field definitions in a WIT, despite the API returning the correct result. The restriction imposed for correctness is relaxed and changed into an Assert, so this condition can be caught in Debug builds. Since Release builds are distributed to customers, the mapping will not be added and future invocations will refer to the existing element.

That is, if IterationPath is contained in ordinal positions 0, 48 the addition of the field at ordinal 48 will not be performed. Any future reference to IterationPath will refer to the field received at position 0.